Refactor Guest actions hpp=>h/cpp
