+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -22,8 +22,8 @@
 #include "FootpathPlaceFromTrackAction.h"
 #include "FootpathRemoveAction.h"
 #include "GameAction.h"
-#include "GuestSetFlagsAction.hpp"
-#include "GuestSetNameAction.hpp"
+#include "GuestSetFlagsAction.h"
+#include "GuestSetNameAction.h"
 #include "LandBuyRightsAction.hpp"
 #include "LandLowerAction.hpp"
 #include "LandRaiseAction.hpp"
