+++ src/openrct2/actions/GuestSetNameAction.cpp
@@ -7,7 +7,7 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
-#pragma once
+#include "GuestSetNameAction.h"
 
 #include "../Cheats.h"
 #include "../Context.h"
@@ -19,99 +19,68 @@
 #include "../windows/Intent.h"
 #include "../world/Park.h"
 #include "../world/Sprite.h"
-#include "GameAction.h"
 
-DEFINE_GAME_ACTION(GuestSetNameAction, GAME_COMMAND_SET_GUEST_NAME, GameActions::Result)
+void GuestSetNameAction::AcceptParameters(GameActionParameterVisitor& visitor)
 {
-private:
-    uint16_t _spriteIndex{ SPRITE_INDEX_NULL };
-    std::string _name;
+    visitor.Visit("peep", _spriteIndex);
+    visitor.Visit("name", _name);
+}
 
-public:
-    GuestSetNameAction() = default;
-    GuestSetNameAction(uint16_t spriteIndex, const std::string& name)
-        : _spriteIndex(spriteIndex)
-        , _name(name)
+void GuestSetNameAction::Serialise(DataSerialiser& stream)
+{
+    GameAction::Serialise(stream);
+
+    stream << DS_TAG(_spriteIndex) << DS_TAG(_name);
+}
+
+GameActions::Result::Ptr GuestSetNameAction::Query() const
+{
+    if (_spriteIndex >= MAX_SPRITES)
     {
+        return std::make_unique<GameActions::Result>(GameActions::Status::InvalidParameters, STR_CANT_NAME_GUEST, STR_NONE);
     }
 
-    void AcceptParameters(GameActionParameterVisitor & visitor) override
+    auto guest = TryGetEntity<Guest>(_spriteIndex);
+    if (guest == nullptr)
     {
-        visitor.Visit("peep", _spriteIndex);
-        visitor.Visit("name", _name);
+        log_warning("Invalid game command for sprite %u", _spriteIndex);
+        return std::make_unique<GameActions::Result>(GameActions::Status::InvalidParameters, STR_CANT_NAME_GUEST, STR_NONE);
     }
 
-    uint16_t GetSpriteIndex() const
+    return std::make_unique<GameActions::Result>();
+}
+
+GameActions::Result::Ptr GuestSetNameAction::Execute() const
+{
+    auto guest = TryGetEntity<Guest>(_spriteIndex);
+    if (guest == nullptr)
     {
-        return _spriteIndex;
+        log_warning("Invalid game command for sprite %u", _spriteIndex);
+        return std::make_unique<GameActions::Result>(GameActions::Status::InvalidParameters, STR_CANT_NAME_GUEST, STR_NONE);
     }
 
-    std::string GetGuestName() const
+    auto curName = guest->GetName();
+    if (curName == _name)
     {
-        return _name;
+        return std::make_unique<GameActions::Result>(GameActions::Status::Ok, STR_NONE);
     }
 
-    uint16_t GetActionFlags() const override
+    if (!guest->SetName(_name))
     {
-        return GameAction::GetActionFlags() | GameActions::Flags::AllowWhilePaused;
+        return std::make_unique<GameActions::Result>(GameActions::Status::Unknown, STR_CANT_NAME_GUEST, STR_NONE);
     }
 
-    void Serialise(DataSerialiser & stream) override
-    {
-        GameAction::Serialise(stream);
+    // Easter egg functions are for guests only
+    guest->HandleEasterEggName();
 
-        stream << DS_TAG(_spriteIndex) << DS_TAG(_name);
-    }
+    gfx_invalidate_screen();
 
-    GameActions::Result::Ptr Query() const override
-    {
-        if (_spriteIndex >= MAX_SPRITES)
-        {
-            return std::make_unique<GameActions::Result>(GameActions::Status::InvalidParameters, STR_CANT_NAME_GUEST, STR_NONE);
-        }
+    auto intent = Intent(INTENT_ACTION_REFRESH_GUEST_LIST);
+    context_broadcast_intent(&intent);
 
-        auto guest = TryGetEntity<Guest>(_spriteIndex);
-        if (guest == nullptr)
-        {
-            log_warning("Invalid game command for sprite %u", _spriteIndex);
-            return std::make_unique<GameActions::Result>(GameActions::Status::InvalidParameters, STR_CANT_NAME_GUEST, STR_NONE);
-        }
-
-        return std::make_unique<GameActions::Result>();
-    }
-
-    GameActions::Result::Ptr Execute() const override
-    {
-        auto guest = TryGetEntity<Guest>(_spriteIndex);
-        if (guest == nullptr)
-        {
-            log_warning("Invalid game command for sprite %u", _spriteIndex);
-            return std::make_unique<GameActions::Result>(GameActions::Status::InvalidParameters, STR_CANT_NAME_GUEST, STR_NONE);
-        }
-
-        auto curName = guest->GetName();
-        if (curName == _name)
-        {
-            return std::make_unique<GameActions::Result>(GameActions::Status::Ok, STR_NONE);
-        }
-
-        if (!guest->SetName(_name))
-        {
-            return std::make_unique<GameActions::Result>(GameActions::Status::Unknown, STR_CANT_NAME_GUEST, STR_NONE);
-        }
-
-        // Easter egg functions are for guests only
-        guest->HandleEasterEggName();
-
-        gfx_invalidate_screen();
-
-        auto intent = Intent(INTENT_ACTION_REFRESH_GUEST_LIST);
-        context_broadcast_intent(&intent);
-
-        auto res = std::make_unique<GameActions::Result>();
-        res->Position.x = guest->x;
-        res->Position.y = guest->y;
-        res->Position.z = guest->z;
-        return res;
-    }
-};
+    auto res = std::make_unique<GameActions::Result>();
+    res->Position.x = guest->x;
+    res->Position.y = guest->y;
+    res->Position.z = guest->z;
+    return res;
+}
